requirements.txt
anthropic>=0.42.0
httpx>=0.27.0
aiohttp>=3.9.0
web3>=6.15.0
python-dotenv>=1.0.0
rich>=13.7.0
apscheduler>=3.10.0
py-clob-client>=0.34.0
